# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/barbers": {
            "post": {
                "tags": [
                    "Barber"
                ],
                "summary": "Store a new barber.",
                "description": "Store a new barber with name, barbershop id, and image.",
                "operationId": "0d0e917b6d219bc19243f624f4b96c18",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "barbershop_id",
                                    "image"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "The name of the barber.",
                                        "type": "string"
                                    },
                                    "barbershop_id": {
                                        "description": "The ID of the barbershop that the barber belongs to.",
                                        "type": "integer"
                                    },
                                    "image": {
                                        "description": "The image of the barber.",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Barber created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Barber created successfully"
                                        },
                                        "barber": {}
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/barbers/{barber_id}": {
            "put": {
                "tags": [
                    "Barber"
                ],
                "summary": "Update an existing barber.",
                "description": "Update an existing barber with name, barbershop id, and image.",
                "operationId": "8910504062b68f135cc5403051901510",
                "parameters": [
                    {
                        "name": "barber_id",
                        "in": "path",
                        "description": "The ID of the barber to update.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "barbershop_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "The name of the barber.",
                                        "type": "string"
                                    },
                                    "barbershop_id": {
                                        "description": "The ID of the barbershop that the barber belongs to.",
                                        "type": "integer"
                                    },
                                    "image": {
                                        "description": "The image of the barber.",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Barber updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Barber updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Barber not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Barber not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Barber"
                ],
                "summary": "Delete a barber.",
                "description": "Delete a barber by ID.",
                "operationId": "bd2de630acf1401abbc16cada5fcfc74",
                "parameters": [
                    {
                        "name": "barber_id",
                        "in": "path",
                        "description": "ID of the barber to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Barber deleted successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The barber has been deleted!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Barber not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "No barber found to be deleted!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}